Fix librofm auth_token setter on init login. Added an amazon base class, moving most of the audible logic there.
